pyproject.toml
[tool.poetry]
name = "rdf-delta-load-test"
version = "0.1.0"
description = ""
authors = ["Lawon Lewis <lawson@kurrawong.ai>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
rdf-graph-gen = "^1.1.3"
httpx = "^0.28.0"
rdflib = "^7.1.1"
jinja2 = "^3.1.4"
docker = "^7.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
